immutable: refactor into a package
