+++ mediatimestamp/immutable/timerange.py
@@ -1,4 +1,4 @@
-# Copyright 2019 British Broadcasting Corporation
+# Copyright 2020 British Broadcasting Corporation
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -12,754 +12,19 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""This library provides a class TimeOffset which stores an immutable signed time difference value with nanosecond
-precision.
-
-It also provides a class Timestamp which is a descendent of TimeOffset which represents an immutable time offset since
-the epoch (ie. 1970-01-01T00:00:00.000000000Z)
-
-And finally it includes an immutable TimeRange object which stores a range between two Timestamps.
-
-These data types are of use in a number of situations, but particularly for code that will handle PTP timestamps, which
-are normally stored in this fashion.
-
-
-
-Expected Logic for binary operations on timestamps and time offsets:
-
-Timestamp and TimeOffset objects can be added and subtracted. The type of the final result depends upon the order and
-type of the operands
-
-TS + TO = TS
-TS + TS = TS (treats 2nd TS as TO)
-TO + TS = TS
-TO + TO = TO
-
-TS - TO = TS
-TS - TS = TO
-TO - TS = TO (treats TS as TO)
-TO - TO = TO
-
-+= and -= always give the same result type as the first operand
-
-Instances of TimeOffset can always be multiplied or divided by integers and floats and always give
-another TimeOffset as a result.
-
-An instance of Timestamp multiplied by integers or floats will be treated as an instance of TimeOffset
-"""
-
-import calendar
-import time
 import re
-from datetime import datetime
-from dateutil import tz
 from fractions import Fraction
-from numbers import Rational
 
 from typing import Tuple, Union, Optional, cast, Iterator
 
-from .constants import MAX_NANOSEC, MAX_SECONDS, UTC_LEAP
-from .exceptions import TsValueError
+from ..constants import MAX_NANOSEC
+from ..exceptions import TsValueError
 
-__all__ = ["TimeOffset", "Timestamp", "TimeRange"]
+from ._types import RationalTypes
+from .timeoffset import TimeOffset
+from .timestamp import Timestamp
 
-
-RationalTypes = Union[int, Rational]
-
-
-def _parse_seconds_fraction(frac: str) -> int:
-    """ Parse the fraction part of a timestamp seconds, using maximum 9 digits
-    Returns the nanoseconds
-    """
-    ns = 0
-    mult = TimeOffset.MAX_NANOSEC
-    for c in frac:
-        if c < '0' or c > '9' or int(mult) < 1:
-            break
-        mult = mult // 10
-        ns += mult * int(c)
-    return ns
-
-
-def _parse_iso8601(iso8601: str) -> Tuple[int, int, int, int, int, int, int]:
-    """ Limited ISO 8601 timestamp parse; expands YYYY-MM-DDThh:mm:ss.s
-    Returns tuple of (year, month, day, hours, mins, seconds, nanoseconds)
-    """
-    iso_date_time = iso8601.split("T")
-    if len(iso_date_time) != 2:
-        raise TsValueError("invalid or unsupported ISO 8601 UTC format")
-    iso_date = iso_date_time[0].split("-")
-    iso_time = iso_date_time[1].split(":")
-    if len(iso_date) != 3 or len(iso_time) != 3:
-        raise TsValueError("invalid or unsupported ISO 8601 UTC format")
-    sec_frac = iso_time[2].split(".")
-    if len(sec_frac) != 1 and len(sec_frac) != 2:
-        raise TsValueError("invalid or unsupported ISO 8601 UTC format")
-    sec = sec_frac[0]
-    ns = 0
-    if len(sec_frac) > 1:
-        ns = _parse_seconds_fraction(sec_frac[1])
-    return (int(iso_date[0]), int(iso_date[1]), int(iso_date[2]), int(iso_time[0]), int(iso_time[1]), int(sec), ns)
-
-
-TimeOffsetConstructionType = Union["TimeOffset", int, float]
-
-
-class TimeOffset(object):
-    """A nanosecond precision immutable time difference object.
-
-    Note that the canonical representation of a TimeOffset is seconds:nanoseconds, e.g. "4:500000000".
-    TimeOffsets in seconds.fractions format (e.g. "4.5") can be parsed, but should not be used for serialization or
-    storage due to difficulty disambiguating them from floats.
-    """
-    class Rounding (int):
-        pass
-
-    ROUND_DOWN = Rounding(0)
-    ROUND_NEAREST = Rounding(1)
-    ROUND_UP = Rounding(2)
-
-    MAX_NANOSEC = MAX_NANOSEC
-    MAX_SECONDS = MAX_SECONDS
-
-    def __init__(self, sec: int = 0, ns: int = 0, sign: int = 1):
-        (sec, ns, sign) = self._make_valid(int(sec), int(ns), int(sign))
-
-        self.sec: int
-        self.ns: int
-        self.sign: int
-
-        self.__dict__['sec'] = int(sec)
-        self.__dict__['ns'] = int(ns)
-        self.__dict__['sign'] = int(sign)
-
-    def __setattr__(self, name: str, value: object) -> None:
-        raise TsValueError("Cannot assign to an immutable TimeOffset")
-
-    @classmethod
-    def from_timeoffset(cls, toff: "TimeOffset") -> "TimeOffset":
-        return cls(sec=toff.sec, ns=toff.ns, sign=toff.sign)
-
-    @classmethod
-    def get_interval_fraction(cls,
-                              rate_num: RationalTypes,
-                              rate_den: RationalTypes = 1,
-                              factor: int = 1) -> "TimeOffset":
-        if rate_num <= 0 or rate_den <= 0:
-            raise TsValueError("invalid rate")
-        if factor < 1:
-            raise TsValueError("invalid interval factor")
-
-        rate = Fraction(rate_num, rate_den)
-        sec = rate.denominator // (rate.numerator * factor)
-        rem = rate.denominator % (rate.numerator * factor)
-        ns = cls.MAX_NANOSEC * rem // (rate.numerator * factor)
-        return cls(sec=sec, ns=ns)
-
-    @classmethod
-    def from_sec_frac(cls, toff_str: str) -> "TimeOffset":
-        sec_frac = toff_str.split(".")
-        if len(sec_frac) != 1 and len(sec_frac) != 2:
-            raise TsValueError("invalid second.fraction format")
-        sec = int(sec_frac[0])
-        sign = 1
-        if sec_frac[0].startswith("-"):
-            sign = -1
-            sec = abs(sec)
-        ns = 0
-        if len(sec_frac) > 1:
-            ns = _parse_seconds_fraction(sec_frac[1])
-        return cls(sec=sec, ns=ns, sign=sign)
-
-    @classmethod
-    def from_sec_nsec(cls, toff_str: str) -> "TimeOffset":
-        sec_frac = toff_str.split(":")
-        if len(sec_frac) != 1 and len(sec_frac) != 2:
-            raise TsValueError("invalid second:nanosecond format")
-        sec = int(sec_frac[0])
-        sign = 1
-        if sec_frac[0].startswith("-"):
-            sign = -1
-            sec = abs(sec)
-        ns = 0
-        if len(sec_frac) > 1:
-            ns = int(sec_frac[1])
-        return cls(sec=sec, ns=ns, sign=sign)
-
-    @classmethod
-    def from_str(cls, toff_str: str) -> "TimeOffset":
-        """Parse a string as a TimeOffset
-
-        Accepts both second:nanosecond and second.fraction formats.
-        """
-        if '.' in toff_str:
-            return cls.from_sec_frac(toff_str)
-        else:
-            return cls.from_sec_nsec(toff_str)
-
-    @classmethod
-    def from_count(cls, count: int, rate_num: RationalTypes, rate_den: RationalTypes = 1) -> "TimeOffset":
-        """Returns a new TimeOffset derived from a count and a particular rate.
-
-        :param count: The sample count
-        :param rate_num: The numerator of the rate, in Hz
-        :param rate_den: The denominator of the rate in Hz
-        """
-        if rate_num <= 0 or rate_den <= 0:
-            raise TsValueError("invalid rate")
-        rate = Fraction(rate_num, rate_den)
-        abs_count = abs(count)
-        sec = (abs_count * rate.denominator) // rate.numerator
-        rem = (abs_count * rate.denominator) % rate.numerator
-        ns = (rem * cls.MAX_NANOSEC) // rate.numerator
-        sign = 1
-        if count < 0:
-            sign = -1
-        return cls(sec=sec, ns=ns, sign=sign)
-
-    @classmethod
-    def from_millisec(cls, millisec: int) -> "TimeOffset":
-        abs_millisec = abs(millisec)
-        sec = abs_millisec // 1000
-        ns = (abs_millisec % 1000) * 1000000
-        sign = 1
-        if millisec < 0:
-            sign = -1
-        return cls(sec=sec, ns=ns, sign=sign)
-
-    @classmethod
-    def from_microsec(cls, microsec: int) -> "TimeOffset":
-        abs_microsec = abs(microsec)
-        sec = abs_microsec // 1000000
-        ns = (abs_microsec % 1000000) * 1000
-        sign = 1
-        if microsec < 0:
-            sign = -1
-        return cls(sec=sec, ns=ns, sign=sign)
-
-    @classmethod
-    def from_nanosec(cls, nanosec: int) -> "TimeOffset":
-        abs_nanosec = abs(nanosec)
-        sec = abs_nanosec // cls.MAX_NANOSEC
-        ns = abs_nanosec % cls.MAX_NANOSEC
-        sign = 1
-        if nanosec < 0:
-            sign = -1
-        return cls(sec=sec, ns=ns, sign=sign)
-
-    def is_null(self) -> bool:
-        return self.sec == 0 and self.ns == 0
-
-    def to_sec_nsec(self) -> str:
-        """ Convert to <seconds>:<nanoseconds>
-        """
-        strSign = ""
-        if self.sign < 0:
-            strSign = "-"
-        return u"{}{}:{}".format(strSign, self.sec, self.ns)
-
-    def to_sec_frac(self, fixed_size: bool = False) -> str:
-        """ Convert to <seconds>.<fraction>
-        """
-        strSign = ""
-        if self.sign < 0:
-            strSign = "-"
-        return u"{}{}.{}".format(
-            strSign,
-            self.sec,
-            self._get_fractional_seconds(fixed_size=fixed_size))
-
-    def to_count(self, rate_num: RationalTypes, rate_den: RationalTypes = 1,
-                 rounding: "TimeOffset.Rounding" = ROUND_NEAREST) -> int:
-        """Returns an integer such that if this TimeOffset is equal to an exact number of samples at the given rate
-        then this is equal, and otherwise the value is rounded as indicated by the rounding parameter.
-
-        :param rate_num: numerator of rate
-        :param rate_den: denominator of rate
-        :param rounding: One of TimeOffset.ROUND_NEAREST, TimeOffset.ROUND_UP, or TimeOffset.ROUND_DOWN
-        """
-        if rate_num <= 0 or rate_den <= 0:
-            raise TsValueError("invalid rate")
-        rate = Fraction(rate_num, rate_den)
-        abs_off = self.__abs__()
-        use_rounding = rounding
-        if self.sign < 0:
-            if use_rounding == self.ROUND_UP:
-                use_rounding = self.ROUND_DOWN
-            elif use_rounding == self.ROUND_DOWN:
-                use_rounding = self.ROUND_UP
-        if use_rounding == self.ROUND_NEAREST:
-            rnd_off = TimeOffset.get_interval_fraction(rate, factor=2)
-        elif use_rounding == self.ROUND_UP:
-            rnd_off = TimeOffset.get_interval_fraction(rate, factor=1) - TimeOffset(0, 1)
-        else:
-            rnd_off = TimeOffset()
-        if rnd_off.sign > 0:
-            abs_off += rnd_off
-
-        # off_at_rate = (off_sec + off_nsec) / (rate_den / rate_num)
-        #             = (off_sec x rate_num / rate_den) + (off_nsec x rate_num) / rate_den)
-        #             = {f1} + {f2}
-        # reduce {f1} as follows: a*b/c = ((a/c)*c + a%c) * b) / c = (a/c)*b + (a%c)*b/c
-        # then combine the f1 and f2 nanosecond values before division by rate_den to avoid loss of precision when
-        # off_sec and off_nsec are not multiples of rate_den, but (off_sec + off_nsec) is
-        f1_whole = (abs_off.sec // rate.denominator) * rate.numerator
-        f1_dennsec = (abs_off.sec % rate.denominator) * rate.numerator * self.MAX_NANOSEC
-        f2_dennsec = abs_off.ns * rate.numerator
-        return self.sign * (f1_whole + (f1_dennsec + f2_dennsec) // (rate.denominator * self.MAX_NANOSEC))
-
-    def to_phase_offset(self, rate_num: RationalTypes, rate_den: RationalTypes = 1) -> "TimeOffset":
-        """Return the smallest positive TimeOffset such that abs(self - returnval) represents an integer number of
-        samples at the given rate"""
-        return self - self.normalise(rate_num, rate_den, rounding=TimeOffset.ROUND_DOWN)
-
-    def to_millisec(self, rounding: "TimeOffset.Rounding" = ROUND_NEAREST) -> int:
-        use_rounding = rounding
-        if self.sign < 0:
-            if use_rounding == self.ROUND_UP:
-                use_rounding = self.ROUND_DOWN
-            elif use_rounding == self.ROUND_DOWN:
-                use_rounding = self.ROUND_UP
-        round_ns = 0
-        if use_rounding == self.ROUND_NEAREST:
-            round_ns = 1000000 // 2
-        elif use_rounding == self.ROUND_UP:
-            round_ns = 1000000 - 1
-        return self.sign * (self.sec*1000 + (self.ns + round_ns)//1000000)
-
-    def to_microsec(self, rounding: "TimeOffset.Rounding" = ROUND_NEAREST) -> int:
-        use_rounding = rounding
-        if self.sign < 0:
-            if use_rounding == self.ROUND_UP:
-                use_rounding = self.ROUND_DOWN
-            elif use_rounding == self.ROUND_DOWN:
-                use_rounding = self.ROUND_UP
-        round_ns = 0
-        if use_rounding == self.ROUND_NEAREST:
-            round_ns = 1000 // 2
-        elif use_rounding == self.ROUND_UP:
-            round_ns = 1000 - 1
-        return self.sign * (self.sec*1000000 + (self.ns + round_ns)//1000)
-
-    def to_nanosec(self) -> int:
-        return self.sign * (self.sec*self.MAX_NANOSEC + self.ns)
-
-    def normalise(self,
-                  rate_num: RationalTypes,
-                  rate_den: RationalTypes = 1,
-                  rounding: "TimeOffset.Rounding" = ROUND_NEAREST) -> "TimeOffset":
-        """Return the nearest TimeOffset to self which represents an integer number of samples at the given rate.
-
-        :param rate_num: Rate numerator
-        :param rate_den: Rate denominator
-        :param rounding: How to round, if set to TimeOffset.ROUND_DOWN (resp. TimeOffset.ROUND_UP) this method will only
-                         return a TimeOffset less than or equal to this one (resp. greater than or equal to).
-        """
-        return self.from_count(self.to_count(rate_num, rate_den, rounding), rate_num, rate_den)
-
-    def compare(self, other_in: TimeOffsetConstructionType) -> int:
-        other = self._cast_arg(other_in)
-        if self.sign != other.sign:
-            return self.sign
-        elif self.sec < other.sec:
-            return -self.sign
-        elif self.sec > other.sec:
-            return self.sign
-        elif self.ns < other.ns:
-            return -self.sign
-        elif self.ns > other.ns:
-            return self.sign
-        else:
-            return 0
-
-    def __str__(self) -> str:
-        return self.to_sec_nsec()
-
-    def __repr__(self) -> str:
-        return "{}.from_sec_nsec({!r})".format(type(self).__module__ + '.' + type(self).__name__, self.to_sec_nsec())
-
-    def __abs__(self) -> "TimeOffset":
-        return TimeOffset(self.sec, self.ns, 1)
-
-    def __hash__(self) -> int:
-        return self.to_nanosec()
-
-    def __eq__(self, other: object) -> bool:
-        return isinstance(other, (int, float, TimeOffset)) and self.compare(other) == 0
-
-    def __ne__(self, other: object) -> bool:
-        return not (self == other)
-
-    def __lt__(self, other: TimeOffsetConstructionType) -> bool:
-        return self.compare(other) < 0
-
-    def __le__(self, other: TimeOffsetConstructionType) -> bool:
-        return self.compare(other) <= 0
-
-    def __gt__(self, other: TimeOffsetConstructionType) -> bool:
-        return self.compare(other) > 0
-
-    def __ge__(self, other: TimeOffsetConstructionType) -> bool:
-        return self.compare(other) >= 0
-
-    def __add__(self, other_in: TimeOffsetConstructionType) -> "TimeOffset":
-        other = self._cast_arg(other_in)
-        sec = self.sign*self.sec + other.sign*other.sec
-        ns = self.sign*self.ns + other.sign*other.ns
-
-        if not isinstance(self, Timestamp) and not isinstance(other, Timestamp):
-            return TimeOffset(sec, ns)
-        else:
-            return Timestamp(sec, ns)
-
-    def __sub__(self, other_in: TimeOffsetConstructionType) -> "TimeOffset":
-        other = self._cast_arg(other_in)
-        sec = self.sign*self.sec - other.sign*other.sec
-        ns = self.sign*self.ns - other.sign*other.ns
-
-        if isinstance(self, Timestamp) and not isinstance(other, Timestamp):
-            return Timestamp(sec, ns)
-        else:
-            return TimeOffset(sec, ns)
-
-    def __iadd__(self, other_in: TimeOffsetConstructionType) -> "TimeOffset":
-        other = self._cast_arg(other_in)
-        tmp = self + other
-        return self.__class__(tmp.sec, tmp.ns, tmp.sign)
-
-    def __isub__(self, other_in: TimeOffsetConstructionType) -> "TimeOffset":
-        other = self._cast_arg(other_in)
-        tmp = self - other
-        return self.__class__(tmp.sec, tmp.ns, tmp.sign)
-
-    def __mul__(self, anint: int) -> "TimeOffset":
-        sec = self.sec * abs(anint)
-        ns = self.ns * abs(anint)
-
-        if anint < 0:
-            sign = self.sign * -1
-        else:
-            sign = self.sign
-
-        if ns >= self.MAX_NANOSEC:
-            sec += (ns // self.MAX_NANOSEC)
-            ns %= self.MAX_NANOSEC
-
-        return TimeOffset(sec, ns, sign)
-
-    def __rmul__(self, anint: int) -> "TimeOffset":
-        return (self * anint)
-
-    def __div__(self, anint: int) -> "TimeOffset":
-        return (self // anint)
-
-    def __truediv__(self, anint: int) -> "TimeOffset":
-        return (self // anint)
-
-    def __floordiv__(self, anint: int) -> "TimeOffset":
-        (sec, ns, sign) = (self.sec, self.ns, self.sign)
-        abs_anint = abs(anint)
-        sec = sec // abs_anint
-        ns = int((self.ns + (self.sec % abs_anint) * self.MAX_NANOSEC) / abs_anint + 5e-10)
-
-        sec = sec + ns // self.MAX_NANOSEC
-        ns = ns % self.MAX_NANOSEC
-        if anint < 0:
-            sign *= -1
-        return TimeOffset(sec, ns, sign)
-
-    def _get_fractional_seconds(self, fixed_size: bool = False) -> str:
-        div = self.MAX_NANOSEC // 10
-        rem = self.ns
-        sec_frac = ""
-
-        for i in range(0, 9):
-            if not fixed_size and i > 0 and rem == 0:
-                break
-
-            sec_frac += '%i' % (rem / div)
-            rem %= div
-            div //= 10
-
-        return sec_frac
-
-    def _cast_arg(self, other: TimeOffsetConstructionType) -> "TimeOffset":
-        if isinstance(other, int):
-            return TimeOffset(other)
-        elif isinstance(other, float):
-            return TimeOffset.from_sec_frac(str(other))
-        elif isinstance(other, TimeOffset) and not isinstance(other, TimeOffset):
-            return TimeOffset.from_timeoffset(other)
-        else:
-            return other
-
-    def _make_valid(self, sec: int, ns: int, sign: int) -> Tuple[int, int, int]:
-        if sign > 0 or (sec == 0 and ns == 0):
-            sign = 1
-        else:
-            sign = -1
-
-        sec += (ns // self.MAX_NANOSEC)
-        ns %= self.MAX_NANOSEC
-
-        if sec < 0 or (sec == 0 and ns < 0):
-            sec *= -1
-            ns *= -1
-            sign *= -1
-
-        if ns < 0:
-            sec -= 1
-            ns += self.MAX_NANOSEC
-
-        if sec >= self.MAX_SECONDS:
-            sec = self.MAX_SECONDS - 1
-            ns = self.MAX_NANOSEC - 1
-
-        return (sec, ns, sign)
-
-
-class Timestamp(TimeOffset):
-    """A nanosecond precision immutable timestamp."""
-    def __init__(self, sec: int = 0, ns: int = 0, sign: int = 1):
-        super(Timestamp, self).__init__(sec, ns, sign)
-
-    def __setattr__(self, name: str, value: object) -> None:
-        raise TsValueError("Cannot assign to an immutable Timestamp")
-
-    @classmethod
-    def get_time(cls, *, force_pure_python=False) -> "Timestamp":
-        """The force_pure_python keyword only argument is ignored."""
-        utc_time = time.time()
-        return cls.from_utc(int(utc_time), int(utc_time*cls.MAX_NANOSEC) - int(utc_time)*cls.MAX_NANOSEC)
-
-    @classmethod
-    def from_timeoffset(cls, toff: TimeOffset) -> "Timestamp":
-        return cls(sec=toff.sec, ns=toff.ns, sign=toff.sign)
-
-    @classmethod
-    def from_sec_frac(cls, ts_str: str) -> "Timestamp":
-        return cast(Timestamp, super(Timestamp, cls).from_sec_frac(ts_str))
-
-    @classmethod
-    def from_tai_sec_frac(cls, ts_str: str) -> "Timestamp":
-        return cls.from_sec_frac(ts_str)
-
-    @classmethod
-    def from_sec_nsec(cls, ts_str: str) -> "Timestamp":
-        return cast(Timestamp, super(Timestamp, cls).from_sec_nsec(ts_str))
-
-    @classmethod
-    def from_tai_sec_nsec(cls, ts_str: str) -> "Timestamp":
-        return cls.from_sec_nsec(ts_str)
-
-    @classmethod
-    def from_datetime(cls, dt: datetime) -> "Timestamp":
-        minTs = datetime.fromtimestamp(0, tz.gettz('UTC'))
-        utcdt = dt.astimezone(tz.gettz('UTC'))
-        seconds = int((utcdt - minTs).total_seconds())
-        nanoseconds = utcdt.microsecond * 1000
-
-        return cls.from_utc(seconds, nanoseconds, False)
-
-    @classmethod
-    def from_iso8601_utc(cls, iso8601utc: str) -> "Timestamp":
-        if not iso8601utc.endswith('Z'):
-            raise TsValueError("missing 'Z' at end of ISO 8601 UTC format")
-        year, month, day, hour, minute, second, ns = _parse_iso8601(iso8601utc[:-1])
-        gmtuple = (year, month, day, hour, minute, second - (second == 60))
-        secs_since_epoch = calendar.timegm(gmtuple)
-        return cls.from_utc(secs_since_epoch, ns, (second == 60))
-
-    @classmethod
-    def from_smpte_timelabel(cls, timelabel: str) -> "Timestamp":
-        r = re.compile(r'(\d+)-(\d+)-(\d+)T(\d+):(\d+):(\d+)F(\d+) (\d+)/(\d+) UTC([-\+])(\d+):(\d+) TAI([-\+])(\d+)')
-        m = r.match(timelabel)
-        if m is None:
-            raise TsValueError("invalid SMPTE Time Label string format")
-        groups = m.groups()
-        leap_sec = int(int(groups[5]) == 60)
-        local_tm_sec = calendar.timegm(time.struct_time((int(groups[0]), int(groups[1]), int(groups[2]),
-                                                         int(groups[3]), int(groups[4]), int(groups[5]) - leap_sec,
-                                                         0, 0, 0)))
-        rate_num = int(groups[7])
-        rate_den = int(groups[8])
-        utc_sign = 1
-        if groups[9] == '-':
-            utc_sign = -1
-        tai_sign = 1
-        if groups[12] == '-':
-            tai_sign = -1
-        tai_seconds = (local_tm_sec +
-                       leap_sec -
-                       utc_sign*(int(groups[10])*60*60 + int(groups[11])*60) -
-                       tai_sign*int(groups[13]))
-        count = Timestamp(tai_seconds, 0).to_count(rate_num, rate_den, cls.ROUND_UP)
-        count += int(groups[6])
-        return cls.from_count(count, rate_num, rate_den)
-
-    @classmethod
-    def from_str(cls, ts_str: str, *, force_pure_python=False) -> "Timestamp":
-        """Parse a string as a TimeStamp
-
-        Accepts SMPTE timelabel, ISO 8601 UTC, second:nanosecond and second.fraction formats, along with "now" to mean
-        the current time.
-
-        The force_pure_python keyword only argument is ignored.
-        """
-        if 'F' in ts_str:
-            return cls.from_smpte_timelabel(ts_str)
-        elif 'T' in ts_str:
-            return cls.from_iso8601_utc(ts_str)
-        elif ts_str.strip() == 'now':
-            return cls.get_time()
-        else:
-            return cast(Timestamp, super(Timestamp, cls).from_str(ts_str))
-
-    @classmethod
-    def from_count(cls, count: int, rate_num: RationalTypes, rate_den: RationalTypes = 1) -> "Timestamp":
-        return cast(Timestamp, super(Timestamp, cls).from_count(count, rate_num, rate_den))
-
-    @classmethod
-    def from_utc(cls, utc_sec: int, utc_ns: int, is_leap: bool = False) -> "Timestamp":
-        leap_sec = 0
-        for tbl_sec, tbl_tai_sec_minus_1 in UTC_LEAP:
-            if utc_sec >= tbl_sec:
-                leap_sec = (tbl_tai_sec_minus_1 + 1) - tbl_sec
-                break
-        return cls(sec=utc_sec+leap_sec+is_leap, ns=utc_ns)
-
-    def get_leap_seconds(self) -> int:
-        """ Get the UTC leaps seconds.
-        Returns the number of leap seconds that the timestamp is adjusted by when
-        converting to UTC.
-        """
-        leap_sec = 0
-        for utc_sec, tai_sec_minus_1 in UTC_LEAP:
-            if self.sec >= tai_sec_minus_1:
-                leap_sec = (tai_sec_minus_1 + 1) - utc_sec
-                break
-
-        return leap_sec
-
-    def to_tai_sec_nsec(self) -> str:
-        return self.to_sec_nsec()
-
-    def to_tai_sec_frac(self, fixed_size: bool = False) -> str:
-        return self.to_sec_frac(fixed_size=fixed_size)
-
-    def to_datetime(self) -> datetime:
-        sec, nsec, leap = self.to_utc()
-        dt = datetime.fromtimestamp(sec, tz.gettz('UTC'))
-        dt = dt.replace(microsecond=int(round(nsec/1000)))
-
-        return dt
-
-    def to_utc(self) -> Tuple[int, int, bool]:
-        """ Convert to UTC.
-        Returns a tuple of (seconds, nanoseconds, is_leap), where `is_leap` is
-        `True` when the input time corresponds exactly to a UTC leap second.
-        Note that this deliberately returns a tuple, to try and avoid confusion.
-        """
-        leap_sec = 0
-        is_leap = False
-        for utc_sec, tai_sec_minus_1 in UTC_LEAP:
-            if self.sec >= tai_sec_minus_1:
-                leap_sec = (tai_sec_minus_1 + 1) - utc_sec
-                is_leap = self.sec == tai_sec_minus_1
-                break
-
-        return (self.sec - leap_sec, self.ns, is_leap)
-
-    def to_iso8601_utc(self) -> str:
-        """ Get printed representation in ISO8601 format (UTC)
-        YYYY-MM-DDThh:mm:ss.s
-        where `s` is fractional seconds at nanosecond precision (always 9-chars wide)
-        """
-        utc_s, utc_ns, is_leap = self.to_utc()
-        utc_bd = time.gmtime(utc_s)
-        frac_sec = self._get_fractional_seconds(fixed_size=True)
-        leap_sec = int(is_leap)
-        return '%04d-%02d-%02dT%02d:%02d:%02d.%sZ' % (utc_bd.tm_year,
-                                                      utc_bd.tm_mon,
-                                                      utc_bd.tm_mday,
-                                                      utc_bd.tm_hour,
-                                                      utc_bd.tm_min,
-                                                      utc_bd.tm_sec + leap_sec,
-                                                      frac_sec)
-
-    def to_smpte_timelabel(self,
-                           rate_num: RationalTypes,
-                           rate_den: RationalTypes = 1,
-                           utc_offset: Optional[int] = None) -> str:
-        if rate_num <= 0 or rate_den <= 0:
-            raise TsValueError("invalid rate")
-        rate = Fraction(rate_num, rate_den)
-        count = self.to_count(rate)
-        normalised_ts = Timestamp.from_count(count, rate)
-        tai_seconds = normalised_ts.sec
-        count_on_or_after_second = Timestamp(tai_seconds, 0).to_count(rate, rounding=self.ROUND_UP)
-        count_within_second = count - count_on_or_after_second
-
-        utc_sec, utc_ns, is_leap = normalised_ts.to_utc()
-        leap_sec = int(is_leap)
-
-        if utc_offset is None:
-            # calculate local time offset
-            utc_offset_sec = time.timezone
-            lt = time.localtime(utc_sec)
-            if lt.tm_isdst > 0:
-                utc_offset_sec += 60*60
-        else:
-            utc_offset_sec = utc_offset
-        utc_offset_sec_abs = abs(utc_offset_sec)
-        utc_offset_hour = utc_offset_sec_abs // (60*60)
-        utc_offset_min = (utc_offset_sec_abs % (60*60)) // 60
-        utc_sign_char = '+'
-        if utc_offset_sec < 0:
-            utc_sign_char = '-'
-
-        utc_bd = time.gmtime(utc_sec + utc_offset_sec)
-
-        tai_offset = utc_sec + leap_sec - tai_seconds
-        tai_sign_char = '+'
-        if tai_offset < 0:
-            tai_sign_char = '-'
-
-        return '%04d-%02d-%02dT%02d:%02d:%02dF%02u %u/%u UTC%c%02u:%02u TAI%c%u' % (
-                    utc_bd.tm_year, utc_bd.tm_mon, utc_bd.tm_mday,
-                    utc_bd.tm_hour, utc_bd.tm_min, utc_bd.tm_sec + leap_sec,
-                    count_within_second,
-                    rate.numerator, rate.denominator,
-                    utc_sign_char, utc_offset_hour, utc_offset_min,
-                    tai_sign_char, abs(tai_offset))
-
-    def normalise(self,
-                  rate_num: RationalTypes,
-                  rate_den: RationalTypes = 1,
-                  rounding: "TimeOffset.Rounding" = TimeOffset.ROUND_NEAREST) -> "Timestamp":
-        """Return the nearest Timestamp to self which represents an integer number of samples at the given rate.
-
-        :param rate_num: Rate numerator
-        :param rate_den: Rate denominator
-        :param rounding: How to round, if set to TimeOffset.ROUND_DOWN (resp. TimeOffset.ROUND_UP) this method will only
-                         return a TimeOffset less than or equal to this one (resp. greater than or equal to).
-        """
-        return self.from_count(self.to_count(rate_num, rate_den, rounding), rate_num, rate_den)
-
-    def __add__(self, other_in: TimeOffsetConstructionType) -> "Timestamp":
-        return cast(Timestamp, super().__add__(other_in))
-
-    def __iadd__(self, other_in: TimeOffsetConstructionType) -> "Timestamp":
-        return cast(Timestamp, super().__add__(other_in))
-
-    def __isub__(self, other_in: TimeOffsetConstructionType) -> "Timestamp":
-        return cast(Timestamp, super().__isub__(other_in))
-
-    def __mul__(self, anint: int) -> "Timestamp":
-        return cast(Timestamp, super().__mul__(anint))
-
-    def __rmul__(self, anint: int) -> "Timestamp":
-        return cast(Timestamp, super().__rmul__(anint))
+__all__ = ["TimeRange"]
 
 
 class TimeRange (object):
@@ -1077,7 +342,7 @@ class TimeRange (object):
                    (self.inclusivity & TimeRange.INCLUDE_END) == (other.inclusivity & TimeRange.INCLUDE_END))))))
 
     def __repr__(self) -> str:
-        return "{}.{}.from_str('{}')".format(type(self).__module__, type(self).__name__, self.to_sec_nsec_range())
+        return "{}.{}.from_str('{}')".format("mediatimestamp.immutable", type(self).__name__, self.to_sec_nsec_range())
 
     def contains_subrange(self, tr: "TimeRange") -> bool:
         """Returns True if the timerange supplied lies entirely inside this timerange"""
